+++ front-end/legeno-around-here/src/components/posting/ImageInput.js
