+++ front-end/legeno-around-here/src/components/posting/Submit.js
@@ -1,0 +1,16 @@
+import styled from 'styled-components';
+
+import submit from '../../images/submit.png';
+
+const Submit = styled.button`
+  background: url(${submit}) no-repeat;
+  background-size: 30px;
+  width: 30px;
+  height: 30px;
+  margin-right: 8px;
+  margin-left: auto;
+  border: 0;
+  outline: 0;
+`;
+
+export default Submit;
