+++ front-end/legeno-around-here/src/components/posting/TopBarBackground.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const TopBarBackground = styled.div`
+  background: #dddddd;
+  color: black;
+  height: 4rem;
+  font-size: 1.5rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+`;
+
+export default TopBarBackground;
